go link: use external linker when in race mode
As of clang 15.0.3 (via zig v0.10), when building with `race = "on"` on
x86_64 Linux, we observe the following:

    runtime/cgo(.text): relocation target memset not defined

Using an external linker when in race mode fixes that. Previously
`linkmode=external` was set for windows-only race builds, now lets just
apply it indiscriminately.

From my past experience, reporting a Go linker error when the external
linker works has a high chance to get the ticket closed as unactionable;
so it makes sense to just use an external linker, when it works.